requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
PyPDF2
langchain
langchain-community
langchain-text-splitters
langchain-openai
pinecone
python-multipart
